+++ R/addParType.R
@@ -4,7 +4,7 @@
 ## should also add parname or par.name which is the concistent
 ## THETA(1), OMEGA(1,1), SIGMA(1,1) labeling.
 addParType <- function(pars,suffix,add.idx){
-
+    
     col.parameter <- "parameter"
     col.par.type <- "par.type"
     if(missing(suffix)) suffix <- NULL
@@ -23,13 +23,18 @@ addParType <- function(pars,suffix,add.idx){
     pars[grepl("^OMEGA",get(col.parameter)),(col.par.type):="OMEGA"]
     pars[grepl("^SIGMA",get(col.parameter)),(col.par.type):="SIGMA"]
     if(add.idx){
-        pars[get(col.par.type)=="THETA",i:=sub("THETA([0-9]+)","\\1",get(col.parameter))]
-        pars[get(col.par.type)=="OMEGA",i:=sub("OMEGA\\(([0-9]+)\\,([0-9]+)\\)","\\1",get(col.parameter))]
-        pars[get(col.par.type)=="OMEGA",j:=sub("OMEGA\\(([0-9]+)\\,([0-9]+)\\)","\\2",get(col.parameter))]
-        pars[get(col.par.type)=="SIGMA",i:=sub("SIGMA\\(([0-9]+)\\,([0-9]+)\\)","\\1",get(col.parameter))]
-        pars[get(col.par.type)=="SIGMA",j:=sub("SIGMA\\(([0-9]+)\\,([0-9]+)\\)","\\2",get(col.parameter))]
-        cols <- cc(i,j)
-        pars[,(cols):=lapply(.SD,as.integer),.SDcols=cols]
+        pars[get(col.par.type)=="THETA",i:=as.integer(sub("THETA([0-9]+)","\\1",get(col.parameter)))]
+        pars[get(col.par.type)=="OMEGA",i:=as.integer(sub("OMEGA\\(([0-9]+)\\,([0-9]+)\\)","\\1",get(col.parameter)))]
+        pars[get(col.par.type)=="OMEGA",j:=as.integer(sub("OMEGA\\(([0-9]+)\\,([0-9]+)\\)","\\2",get(col.parameter)))]
+        pars[get(col.par.type)=="SIGMA",i:=as.integer(sub("SIGMA\\(([0-9]+)\\,([0-9]+)\\)","\\1",get(col.parameter)))]
+        pars[get(col.par.type)=="SIGMA",j:=as.integer(sub("SIGMA\\(([0-9]+)\\,([0-9]+)\\)","\\2",get(col.parameter)))]
+        ## cols <- cc(i,j)
+        ## pars[,(cols):=lapply(.SD,as.integer),.SDcols=cols]
     }
+## par.name
+    pars[,par.name:=parameter]
+    pars[get(col.par.type)=="THETA",par.name:=sprintf("THETA(%s)",i)]
+    
+    
     pars[]
 }